# feat(tExtRef): Export function to lookup control block from ExtRef