--- a/index.ts
+++ b/index.ts
@@ -41,6 +41,7 @@ export { SubscribeOptions } from "./tExtRef/subscribe.js";
 export { subscribe } from "./tExtRef/subscribe.js";
 export { extRefTypeRestrictions } from "./tExtRef/extRefTypeRestrictions.js";
 export { doesFcdaMeetExtRefRestrictions } from "./tExtRef/doesFcdaMeetExtRefRestrictions.js";
+export { sourceControlBlock } from "./tExtRef/sourceControlBlock.js";
 
 export { Supervision } from "./tLN/canInstantiateSubscriptionSupervision.js";
 export { canInstantiateSubscriptionSupervision } from "./tLN/canInstantiateSubscriptionSupervision.js";