--- a/package.json
+++ b/package.json
@@ -3,7 +3,7 @@
   "description": "Function library supporting SCL editing",
   "license": "Apache-2.0",
   "author": "Jakob Vogelsang",
-  "version": "0.6.13",
+  "version": "0.6.14",
   "type": "module",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",