I rewrote the function, added tests and exported it.

I will be most pleased to take a review.