@danyill 2023-10-13
So perhaps the tests are not quite as comprehensive as they could be, but with the old function I have some test failures (11!).

IMHO this is a better function because it just uses the querySelector without the find which should be more performant.

I am unsure about:

* whether `serviceType` should be required which could simplify this a little more but for now I have allowed it to be missing.
* the way I have handled missing attributes vs. empty attributes (I find the standard very unclear in this regard).

I will be most pleased to make improvements.