@danyill 2023-10-15
Thank you for this most helpful review.

I have a couple of questions and I need to write some more tests (I've created stubs but haven't written) them.

If you have a moment to tell me if it looks like it's heading in the right direction I'll fill out the stubs and add any other obviously missing tests.