@danyill 2023-10-15
Just because I had some time, I had a quick hack to look at performance:

set `this!.slow(0)`;

Run 10 tests with old code and new code.
Imagine failing tests with old code don't matter.

Record times:

Old: 

* mean = 76 ms
* std deviation = 40 ms

New

* mean = 83 ms
* std deviation = 24 ms

This is probably not worth much. Tests are probably not the best performance measure and this was done in a wholly unscientific fashion :wink: 

So I tested on a large scd file iterating over a pile of ExtRefs with both the old and new code:

With 387 ExtRefs, of which 223 have a subscription.

- New code: 3364 ms
- Old code: 8128 ms

So I think we're on the right track.
I expect the old code could perhaps be improved substantially by just fully qualifying the  locations but actually it made it worse (8454 ms)!