@danyill 2023-10-15
> We need to conserved more edge cases though, such as srcPrefix is empty string and the prefix of the logical node is not existing. Same I think is true for inst.
> 
>> For inst, this is required for LN and is set to a fixed value of the empty string for LN0. So what the standard means (in at least this case) by "if missing, no inst", means a inst of "". I'm almost tempted to apply this interpretation to prefix but I think it's safest to do as you suggest.

Nope, I am quite wrong -- this is not really true. There is something subtle here, more in a bit.