@JakobVogelsang 2023-10-15
Sorry, I did not read right away to hold off with the review. I general, I do think that performance for this function matters. I am general, therefore, for this changed approach. It is a pleasure to work with you. Doing performance test just because you care. LOVE IT!