@danyill 2023-10-15
OK, clarity emerges:

> We need to conserved more edge cases though, such as srcPrefix is empty string and the prefix of the logical node is not existing. Same I think is true for inst.
> 
>> For inst, this is required for LN and is set to a fixed value of the empty string for LN0. So what the standard means (in at least this case) by "if missing, no inst", means a inst of "". I'm almost tempted to apply this interpretation to prefix but I think it's safest to do as you suggest.
>> 
>>>
>>>Nope, I am quite wrong -- this is not really true. There is something subtle here, more in a bit.
>>>
>>>Perhaps best to hold off review for a bit and I will finish in the next < 24 hrs.

OK, for `inst` I think this is where I'm at:

* For `LN0` `inst` must be set to exactly `""`. It may not be a missing attribute
* For `LN` `inst` must be set to a value and may not be an empty string.

Same applies to `srcLNInst`

With `prefix`, it is the odd one out.

* For `LN0` no prefix is allowed
* For `LN` it is optional and so it could be missing or the empty string so we must check both.