@danyill 2023-10-15
My prefix checking logic is giving me quite a speed up so while it looks like it might not be necessary, for LNs (only) and those with without a srcPrefix we must do twice as much work so it's worth checking for them.

```ts
if (anyLN === "LN") {
    lnPrefixQualifiers =
      srcPrefix && srcPrefix !== ""
        ? [`[prefix="${srcPrefix}"]`]
        : [":not([prefix])", '[prefix=""]'];
  } else {
    lnPrefixQualifiers = [":not([prefix])"];
  }
```