@danyill 2023-10-15
I have corrected a bit of my thinking (this referencing is somewhat inconsistent in its approach and semantics) and I think I've covered a reasonable set of courses.

I shall be delighted for further review and improvement suggestions.